feat: minor improvements to Ethereum delegated siggys
