+++ venus-shared/actors/params.go
@@ -1,5 +1,18 @@
 package actors
 
-import "github.com/filecoin-project/venus/venus-shared/actors/types"
+import (
+	"bytes"
 
-var SerializeParams = types.SerializeParams
+	"github.com/filecoin-project/go-state-types/exitcode"
+	"github.com/filecoin-project/venus/venus-shared/actors/aerrors"
+	cbg "github.com/whyrusleeping/cbor-gen"
+)
+
+func SerializeParams(i cbg.CBORMarshaler) ([]byte, aerrors.ActorError) {
+	buf := new(bytes.Buffer)
+	if err := i.MarshalCBOR(buf); err != nil {
+		// TODO: shouldnt this be a fatal error?
+		return nil, aerrors.Absorb(err, exitcode.ErrSerialization, "failed to encode parameter")
+	}
+	return buf.Bytes(), nil
+}
