+++ venus-shared/actors/types/param.go
@@ -1,14 +1,9 @@
 package types
 
 import (
-	"bytes"
-
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/big"
-	"github.com/filecoin-project/go-state-types/exitcode"
-	"github.com/filecoin-project/venus/venus-shared/actors/aerrors"
 	"github.com/filecoin-project/venus/venus-shared/types/params"
-	cbg "github.com/whyrusleeping/cbor-gen"
 )
 
 var bigZero = big.Zero()
@@ -25,12 +20,3 @@ var ZeroAddress = func() address.Address {
 
 	return ret
 }()
-
-func SerializeParams(i cbg.CBORMarshaler) ([]byte, aerrors.ActorError) {
-	buf := new(bytes.Buffer)
-	if err := i.MarshalCBOR(buf); err != nil {
-		// TODO: shouldnt this be a fatal error?
-		return nil, aerrors.Absorb(err, exitcode.ErrSerialization, "failed to encode parameter")
-	}
-	return buf.Bytes(), nil
-}
